requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.7
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
django-storages==1.14.2
boto3==1.34.0
